Refactor astrometry funcs from utils to AMx
This makes it easier to trace how AMx is calculated.
